fix(android): 移除 glMapBufferRange 中多余的 .toLong()，修复 Kotlin 类型不匹配编译错误
Android GLES30.glMapBufferRange 的 Java 绑定签名为：
  glMapBufferRange(int target, int offset, int length, int access)
length 参数类型为 Int，而非 Long。

上一个 commit（1fa8c55）将 pixelBytes 从 Long 改为 Int 是正确的，
但同时添加了 .toLong() 调用，导致类型不匹配：
  actual type is 'kotlin.Long', but 'kotlin.Int' was expected

修复：移除 pixelBytes.toLong() 中的 .toLong()，直接传入 Int 类型的 pixelBytes。

+++ flutter_app/android/app/src/main/kotlin/com/retrocam/app/camera/CaptureGLProcessor.kt
@@ -747,7 +747,7 @@ void main() {
 
             // 7. Map PBO 读取像素（此时 GPU 传输已完成）
             val mappedBuf = GLES30.glMapBufferRange(
-                GLES30.GL_PIXEL_PACK_BUFFER, 0, pixelBytes.toLong(),
+                GLES30.GL_PIXEL_PACK_BUFFER, 0, pixelBytes,
                 GLES30.GL_MAP_READ_BIT
             ) as? java.nio.ByteBuffer
